+++ components/analyticsComponents/AnalyticsContainer.tsx
@@ -2,25 +2,11 @@ import React from 'https://esm.sh/react';
 import TabBar from './tabs/TabBar.tsx';
 import GraphContainer from './graphs/GraphContainer.tsx';
 import useViewController from '../hooks/useViewController.ts';
-import { useEffect, useState } from 'https://esm.sh/react';
-import calculateMetrics from '../../functions/calculateMetrics.ts';
-import { Result, Snapshot } from '../typings/data.d.ts';
+import useData from "../hooks/useData.ts";
 import '../../style/analyticsContainer.css';
 
 const AnalyticsContainer = () => {
-  const [snapshotArray, setSnapshotArray] = useState<[Snapshot] | null>(null);
-  const [aggregateMetrics, setAggregateMetrics] = useState<Result | null>(null);
-
-  useEffect(() => {
-    fetch('http://localhost:4015/artemis')
-      .then((response) => response.json())
-      .then((data) => {
-        setSnapshotArray(data.artemis);
-        const result: Result = calculateMetrics(data.artemis);
-        setAggregateMetrics(result);
-      })
-      .catch((err) => console.error('UseEffect error', err));
-  }, []);
+  const [snapshotArray, aggregateMetrics] = useData();
   const [view, setView] = useViewController();
 
   return (
